Refactor common code in aws-plugin to aws-common
