+++ data-prepper-plugins/aws-common/build.gradle
@@ -1,7 +1,6 @@
 dependencies {
     implementation project(':data-prepper-api')
     implementation project(':data-prepper-plugins:aws-plugin-api')
-    implementation project(':data-prepper-plugins:aws-common')
     implementation 'com.fasterxml.jackson.core:jackson-databind'
     implementation 'io.micrometer:micrometer-core'
     implementation 'software.amazon.awssdk:auth'
@@ -30,3 +29,4 @@ jacocoTestCoverageVerification {
 }
 
 check.dependsOn jacocoTestCoverageVerification
+
