+++ data-prepper-plugins/aws-common/src/main/java/org/opensearch/dataprepper/aws/common/AwsExtensionProvider.java
@@ -3,17 +3,17 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.dataprepper.plugins.aws;
+package org.opensearch.dataprepper.aws.common;
 
 import org.opensearch.dataprepper.aws.api.AwsCredentialsSupplier;
 import org.opensearch.dataprepper.model.plugin.ExtensionProvider;
 
 import java.util.Optional;
 
-class AwsExtensionProvider implements ExtensionProvider<AwsCredentialsSupplier> {
+public class AwsExtensionProvider implements ExtensionProvider<AwsCredentialsSupplier> {
     private final AwsCredentialsSupplier awsCredentialsSupplier;
 
-    AwsExtensionProvider(final AwsCredentialsSupplier awsCredentialsSupplier) {
+    public AwsExtensionProvider(final AwsCredentialsSupplier awsCredentialsSupplier) {
         this.awsCredentialsSupplier = awsCredentialsSupplier;
     }
