+++ data-prepper-plugins/aws-common/src/main/java/org/opensearch/dataprepper/aws/common/CredentialsCache.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.dataprepper.plugins.aws;
+package org.opensearch.dataprepper.aws.common;
 
 import org.opensearch.dataprepper.aws.api.AwsCredentialsOptions;
 import software.amazon.awssdk.auth.credentials.AwsCredentialsProvider;
@@ -12,14 +12,14 @@ import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Supplier;
 
-class CredentialsCache {
+public class CredentialsCache {
     private final Map<CredentialsIdentifier, AwsCredentialsProvider> credentialsProviderMap;
 
-    CredentialsCache() {
+    public CredentialsCache() {
         credentialsProviderMap = new ConcurrentHashMap<>();
     }
 
-    AwsCredentialsProvider getOrCreate(final AwsCredentialsOptions awsCredentialsOptions, final Supplier<AwsCredentialsProvider> providerSupplier) {
+    public AwsCredentialsProvider getOrCreate(final AwsCredentialsOptions awsCredentialsOptions, final Supplier<AwsCredentialsProvider> providerSupplier) {
         final CredentialsIdentifier identifier = CredentialsIdentifier.fromAwsCredentialsOption(awsCredentialsOptions);
 
         return credentialsProviderMap.computeIfAbsent(identifier, i -> providerSupplier.get());
