+++ data-prepper-plugins/aws-common/src/test/java/org/opensearch/dataprepper/aws/common/AwsExtensionProviderTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.dataprepper.plugins.aws;
+package org.opensearch.dataprepper.aws.common;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
