+++ data-prepper-plugins/aws-plugin/src/test/java/org/opensearch/dataprepper/plugins/aws/DefaultAwsCredentialsSupplierTest.java
@@ -17,6 +17,9 @@ import org.opensearch.dataprepper.aws.api.AwsCredentialsSupplier;
 import software.amazon.awssdk.auth.credentials.AwsCredentialsProvider;
 import software.amazon.awssdk.regions.Region;
 
+import org.opensearch.dataprepper.aws.common.CredentialsProviderFactory;
+import org.opensearch.dataprepper.aws.common.CredentialsCache;
+
 import java.util.List;
 import java.util.Optional;
 import java.util.function.Supplier;
